requirements.txt
tensorflow==1.14.0
keras==2.1.1
imgaug==0.2.6
opencv-python
Pillow
requests
tqdm
sklearn
pytest-cov
codecov
matplotlib
torch==1.3.0
bokeh
sklearn
numpy
glob3
h5py
torchvision
